+++ freemarker/src/main/java/org/jdbi/v3/freemarker/FreemarkerEngine.java
@@ -15,43 +15,35 @@ package org.jdbi.v3.freemarker;
 
 import java.io.IOException;
 import java.io.StringWriter;
+import java.util.Optional;
 import java.util.function.Function;
 
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import org.jdbi.v3.core.config.ConfigRegistry;
-import org.jdbi.v3.core.config.JdbiCache;
-import org.jdbi.v3.core.config.JdbiCaches;
 import org.jdbi.v3.core.statement.StatementContext;
 import org.jdbi.v3.core.statement.TemplateEngine;
 
 /**
  * Rewrites a Freemarker template, using the attributes on the {@link StatementContext} as template parameters.
  */
-public class FreemarkerEngine implements TemplateEngine {
-    private static final JdbiCache<String, Function<StatementContext, String>> RENDER_CACHE =
-            JdbiCaches.declare(FreemarkerEngine::prepare);
-
+public class FreemarkerEngine implements TemplateEngine.Parsing {
     @Override
-    public String render(String sql, StatementContext ctx) {
-        return RENDER_CACHE.get(sql, ctx).apply(ctx);
-    }
-
-    private static Function<StatementContext, String> prepare(ConfigRegistry config, String sql) {
+    public Optional<Function<StatementContext, String>> parse(String sqlTemplate, ConfigRegistry config) {
         try {
-            Template template = new Template(null, sql,
+            Template template = new Template(null, sqlTemplate,
                     config.get(FreemarkerConfig.class).getFreemarkerConfiguration());
-            return ctx -> {
+            return Optional.of(ctx -> {
                 try {
                     StringWriter writer = new StringWriter();
                     template.process(ctx.getAttributes(), writer);
                     return writer.toString();
                 } catch (IOException | TemplateException e) {
-                    throw new IllegalStateException("Failed to render template: " + sql, e);
+                    throw new IllegalStateException("Failed to render template: " + sqlTemplate, e);
                 }
-            };
+            });
         } catch (IOException e) {
-            throw new IllegalStateException("Failed to process template: " + sql, e);
+            throw new IllegalStateException("Failed to process template: " + sqlTemplate, e);
         }
     }
 }